requirements.txt
python-dotenv
mcp
langchain-mcp-adapters
langgraph
langchain-openai
langchain-groq